+++ rs/backend/Cargo.toml
@@ -15,6 +15,8 @@ serde = "1.0.136"
 serde_bytes = "0.11.5"
 serde_cbor = "0.11.2"
 sha2 = "0.10.6"
+strum = "0.24.1"
+strum_macros = "0.24.3"
 lzma-rs = "0.2.0"
 tar = "0.4.37"
 hex = "0.4.3"
