+++ rs/backend/src/accounts_store/schema.rs
@@ -1,9 +1,13 @@
 //! Data storage schemas.
-use crate::accounts_store::Account;
-use ic_crypto_sha::Sha256;
 
+// Schemas
 pub mod map;
 pub mod proxy;
+
+// Mechanics
+use crate::accounts_store::Account;
+use strum_macros::EnumIter;
+pub mod label_serialization;
 #[cfg(test)]
 mod tests;
 
@@ -97,95 +101,15 @@ pub trait AccountsDbTrait {
 ///
 /// Note: The numeric representations of these labels are guaranteed to be stable.
 #[repr(u32)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[derive(Clone, Copy, Debug, EnumIter, PartialEq, Eq)]
 pub enum SchemaLabel {
     /// Data is stored on the heap in a `BTreeMap` and serialized to stable memory on upgrade.
     /// Implemented by: [`map::AccountsDbAsMap`]
     Map = 0,
 }
 
-/// Internal type for the serialized schema label without any kind of safety check.
-pub type SchemaBytesWithoutChecksum = [u8; SchemaLabel::LABEL_BYTES];
 /// Schema Label as written to stable memory.
-type SchemaBytes = [u8; SchemaLabel::MAX_BYTES];
-
-impl TryFrom<u32> for SchemaLabel {
-    type Error = ();
-    fn try_from(value: u32) -> Result<Self, Self::Error> {
-        match value {
-            0 => Ok(Self::Map),
-            _ => Err(()),
-        }
-    }
-}
-
-impl From<SchemaLabel> for SchemaBytesWithoutChecksum {
-    fn from(label: SchemaLabel) -> [u8; SchemaLabel::LABEL_BYTES] {
-        (label as u32).to_le_bytes()
-    }
-}
-
-impl TryFrom<&SchemaBytesWithoutChecksum> for SchemaLabel {
-    type Error = ();
-    fn try_from(value: &[u8; SchemaLabel::LABEL_BYTES]) -> Result<Self, Self::Error> {
-        let label_num = u32::from_le_bytes(*value);
-        Self::try_from(label_num)
-    }
-}
-
-impl From<SchemaLabel> for SchemaBytes {
-    fn from(label: SchemaLabel) -> [u8; SchemaLabel::MAX_BYTES] {
-        let mut bytes = [0u8; SchemaLabel::MAX_BYTES];
-        let label_bytes = SchemaBytesWithoutChecksum::from(label);
-        bytes[SchemaLabel::LABEL_OFFSET..SchemaLabel::LABEL_OFFSET + SchemaLabel::LABEL_BYTES]
-            .copy_from_slice(&label_bytes);
-        let checksum = SchemaLabel::checksum(&label_bytes);
-        bytes[SchemaLabel::CHECKSUM_OFFSET..SchemaLabel::CHECKSUM_OFFSET + SchemaLabel::CHECKSUM_BYTES]
-            .copy_from_slice(&checksum);
-        bytes
-    }
-}
-
-impl TryFrom<&SchemaBytes> for SchemaLabel {
-    type Error = ();
-    fn try_from(value: &[u8; SchemaLabel::MAX_BYTES]) -> Result<Self, Self::Error> {
-        let label_bytes: [u8; SchemaLabel::LABEL_BYTES] = value
-            [SchemaLabel::LABEL_OFFSET..SchemaLabel::LABEL_OFFSET + SchemaLabel::LABEL_BYTES]
-            .try_into()
-            .map_err(|_| ())?; // There are insufficient bytes.
-        let actual_checksum =
-            &value[SchemaLabel::CHECKSUM_OFFSET..SchemaLabel::CHECKSUM_OFFSET + SchemaLabel::CHECKSUM_BYTES];
-        let expected_checksum = Self::checksum(&label_bytes);
-        if expected_checksum == *actual_checksum {
-            Self::try_from(&label_bytes) // Unknown schema label.
-        } else {
-            Err(()) // Invalid checksum.
-        }
-    }
-}
-
-impl SchemaLabel {
-    /// When serialized, the offset of the bytes containing the label.
-    pub const LABEL_OFFSET: usize = 0;
-    /// The number of bytes needed to store just the schema label.
-    pub const LABEL_BYTES: usize = 4;
-    /// When serialized, the offset of the bytes containing the checksum, if included.
-    pub const CHECKSUM_OFFSET: usize = Self::LABEL_BYTES;
-    /// The length of the SHA256 checksum of the schema label.
-    pub const CHECKSUM_BYTES: usize = 32;
-    /// The number of bytes needed to store the schema label and its checksum.
-    pub const MAX_BYTES: usize = Self::CHECKSUM_OFFSET + Self::CHECKSUM_BYTES;
-    /// String used to distinguish this checksum from checksums for other purposes.
-    pub const DOMAIN_SEPARATOR: &[u8; 12] = b"schema-label";
-
-    /// Checksum of the label.
-    fn checksum(self_bytes: &[u8; Self::LABEL_BYTES]) -> [u8; Self::CHECKSUM_BYTES] {
-        let mut state = Sha256::new();
-        state.write(Self::DOMAIN_SEPARATOR);
-        state.write(self_bytes);
-        state.finish()
-    }
-}
+type SchemaLabelBytes = [u8; SchemaLabel::MAX_BYTES];
 
 /// A trait for data stores that support `BTreeMap` for account storage.
 pub trait AccountsDbBTreeMapTrait {
